add waiting kernel

+++ jupyter_server/services/sessions/sessionmanager.py
@@ -8,6 +8,8 @@ import uuid
 from asyncio import Task
 from typing import Any, Dict, List, NewType, Optional, Union, cast
 
+from requests import session
+
 KernelName = NewType("KernelName", str)
 ModelName = NewType("ModelName", str)
 
@@ -429,8 +431,21 @@ class SessionManager(LoggingConfigurable):
         if self.fut_kernel_id_dict is not None and session_id in self.fut_kernel_id_dict:
             model = {
                 "id": session_id,
+                "kernel": {
+                    "id": "waiting",
+                    "name": "Waiting for kernel to start",
+                    "last_activity": "2024-10-02T16:56:56.423328Z",
+                    "execution_state": "waiting",
+                    "connections": 0
+                },
+                "path": "unknown",
+                "type": "notebook",
                 "name": "Waiting for kernel to start",
-                "last_activity": None,
+                "notebook": {
+                    "path": "unknown",
+                    "name": "unknown"
+                },
+                "last_activity": "2024-10-02T16:56:56.423328Z",
                 "execution_state": "waiting",
                 "connections": 0,
             }
@@ -502,8 +517,11 @@ class SessionManager(LoggingConfigurable):
             self.cursor.execute(
                 "SELECT path, name, kernel_id FROM session WHERE session_id=?", [session_id]
             )
-            path, name, kernel_id = self.cursor.fetchone()
-            self.kernel_manager.update_env(kernel_id=kernel_id, env=self.get_kernel_env(path, name))
+            try:
+                path, name, kernel_id = self.cursor.fetchone()
+                self.kernel_manager.update_env(kernel_id=kernel_id, env=self.get_kernel_env(path, name))
+            except TypeError:
+                self.kernel_manager.update_env(kernel_id="waiting", env=self.get_kernel_env("Untitled.ipynb", "Waiting for kernel"))
 
     async def kernel_culled(self, kernel_id: str) -> bool:
         """Checks if the kernel is still considered alive and returns true if its not found."""
@@ -511,28 +529,47 @@ class SessionManager(LoggingConfigurable):
 
     async def row_to_model(self, row, tolerate_culled=False):
         """Takes sqlite database session row and turns it into a dictionary"""
-        kernel_culled: bool = await ensure_async(self.kernel_culled(row["kernel_id"]))
-        if kernel_culled:
-            # The kernel was culled or died without deleting the session.
-            # We can't use delete_session here because that tries to find
-            # and shut down the kernel - so we'll delete the row directly.
-            #
-            # If caller wishes to tolerate culled kernels, log a warning
-            # and return None.  Otherwise, raise KeyError with a similar
-            # message.
-            self.cursor.execute("DELETE FROM session WHERE session_id=?", (row["session_id"],))
-            msg = (
-                "Kernel '{kernel_id}' appears to have been culled or died unexpectedly, "
-                "invalidating session '{session_id}'. The session has been removed.".format(
-                    kernel_id=row["kernel_id"], session_id=row["session_id"]
-                )
-            )
-            if tolerate_culled:
-                self.log.warning(f"{msg}  Continuing...")
-                return None
-            raise KeyError(msg)
 
-        kernel_model = await ensure_async(self.kernel_manager.kernel_model(row["kernel_id"]))
+        dd = dict(row)
+        print(dd)
+        print("-----")
+        kernel_id = row["kernel_id"]
+        if kernel_id == "waiting":
+            if tolerate_culled:
+                return None
+            else:
+                kernel_model = {
+                    "id": "waiting",
+                    "name": "Waiting for kernel to start",
+                    "last_activity": "2024-10-02T16:56:56.423328Z",
+                    "execution_state": "waiting",
+                    "connections": 0,
+                }
+        else:
+            kernel_culled: bool = await ensure_async(self.kernel_culled(kernel_id))
+            if kernel_culled:
+                # The kernel was culled or died without deleting the session.
+                # We can't use delete_session here because that tries to find
+                # and shut down the kernel - so we'll delete the row directly.
+                #
+                # If caller wishes to tolerate culled kernels, log a warning
+                # and return None.  Otherwise, raise KeyError with a similar
+                # message.
+                self.cursor.execute("DELETE FROM session WHERE session_id=?", (row["session_id"],))
+
+                msg = (
+                    "Kernel '{kernel_id}' appears to have been culled or died unexpectedly, "
+                    "invalidating session '{session_id}'. The session has been removed.".format(
+                        kernel_id=row["kernel_id"], session_id=row["session_id"]
+                    )
+                )
+                if tolerate_culled:
+                    self.log.warning(f"{msg}  Continuing...")
+                    return None
+                raise KeyError(msg)
+
+            kernel_model = await ensure_async(self.kernel_manager.kernel_model(kernel_id))
+
         model = {
             "id": row["session_id"],
             "path": row["path"],
@@ -555,7 +592,8 @@ class SessionManager(LoggingConfigurable):
         for row in c.fetchall():
             try:
                 model = await self.row_to_model(row)
-                result.append(model)
+                if model["kernel"]["id"] != "waiting":
+                    result.append(model)
             except KeyError:
                 pass
         return result
